% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{APIkeys}
\alias{APIkeys}
\title{Add API Keys}
\usage{
APIkeys(object, ConsumerKey, SharedSecret)
}
\arguments{
\item{object}{a \code{fatsecret} or \code{fatsecret3L} object}

\item{ConsumerKey}{a character of your FatSecret REST API Consumer Key}

\item{SharedSecret}{a character of your FatSecret REST API Shared Secret}
}
\value{
a \code{fatsecret} or \code{fatsecret3L} object
}
\description{
Add API Keys
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

