% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authToken.R
\name{authToken}
\alias{authToken}
\title{Get Authentication Token (3 Legged - Step 2)}
\usage{
authToken(requestToken)
}
\arguments{
\item{requestToken}{a string of the \code{oauth_token} generate in 3 Legged Authentication Step 1.}

\item{a}{url}
}
\description{
Generate the \code{URL} for User Authorization
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

