% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/AuthoriseUser-method.R
\docType{methods}
\name{AuthoriseUser}
\alias{AuthoriseUser}
\alias{AuthoriseUser,fatsecret3L-method}
\title{Authorise User}
\usage{
AuthoriseUser(object, params)

\S4method{AuthoriseUser}{fatsecret3L}(object, params)
}
\arguments{
\item{object}{a \code{fatsecret3L} object}

\item{params}{a character of the verification code (\code{oauth_verifier})}
}
\value{
a \code{fatsecret3L} object with a valid \code{oauth_token} and \code{oauth_secret} for the user
}
\description{
Authorise User

Retrieve the \code{oauth_token} and \code{oauth_secret} following a successful 3-Legged Authentication request
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
