% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{fatsecret-class}
\alias{fatsecret-class}
\title{fatsecret}
\description{
A class to store essential variables needed for making successful request to the FatSecret RESTful API
}
\section{Slots}{

\describe{
\item{\code{ConsumerKey}}{character string of your FatSecret REST API Comsumer Key}

\item{\code{SharedSecret}}{character string of your FatSecret REST API Shared Secret}

\item{\code{oauth_consumer}}{character of the \code{oauth_consumer} parameter}

\item{\code{RESTURL}}{\code{http://platform.fatsecret.com/rest/server.api}}

\item{\code{httpMethod}}{\code{GET}}

\item{\code{signatureMethod}}{\code{oauth_signature_method=HMAC-SHA1}}

\item{\code{oauthVersion}}{\code{oauth_version=1.0}}

\item{\code{timestamp}}{a character of POSIX timestamp}

\item{\code{nonce}}{a randomly generated 6-digit alpha-numeric string}

\item{\code{methods}}{a \code{data.frame} of currently available methods in \code{fatsecretR}}
}}

\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
