% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{fatsecret3L-class}
\alias{fatsecret3L-class}
\title{fatsecret3L}
\description{
A class to store essential variables needed for making successful 3-Legged Authorisation requests to
existing FatSecret users
}
\section{Slots}{

\describe{
\item{\code{ConsumerKey}}{character string of your FatSecret REST API Comsumer Key}

\item{\code{SharedSecret}}{character string of your FatSecret REST API Shared Secret}

\item{\code{oauth_consumer}}{character of the \code{oauth_consumer} parameter}

\item{\code{RequestURL}}{\code{http://www.fatsecret.com/oauth/request_token}}

\item{\code{AccessURL}}{\code{http://www.fatsecret.com/oauth/access_token}}

\item{\code{AuthURL}}{\code{http://www.fatsecret.com/oauth/authorize}}

\item{\code{httpMethod}}{\code{GET}}

\item{\code{signatureMethod}}{\code{oauth_signature_method=HMAC-SHA1}}

\item{\code{oauthVersion}}{\code{oauth_version=1.0}}

\item{\code{timestamp}}{a character of POSIX timestamp}

\item{\code{nonce}}{a randomly generated 6-digit alpha-numeric string}

\item{\code{oauth_callback}}{the method for authentication callback. If unable to recieve URL callbacks, set to oob. Otherwise
\code{oauth_callback} should be the absolute URL of your callback url}

\item{\code{request_token}}{a character string of the user request token generated in Step-1 of the 3-Legged Authentication process}

\item{\code{request_secret}}{a character string of the user request secret generated in Step-1 of the 3-Legged Authentication process}

\item{\code{user_request_url}}{an authorisation request URL}

\item{\code{user_token}}{the \code{oauth_token} parameter for a user}

\item{\code{user_secret}}{the \code{oauth_secret} parameter for a user}
}}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

