% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/fatsecretR-method.R
\docType{methods}
\name{fatsecretR}
\alias{fatsecretR}
\alias{fatsecretR,fatsecret-method}
\title{fatsecretR REST API}
\usage{
fatsecretR(object, method, param, user_token = NULL, user_secret = NULL)

\S4method{fatsecretR}{fatsecret}(object, method, param, user_token = NULL,
  user_secret = NULL)
}
\arguments{
\item{object}{a \code{fatsecret} object}

\item{method}{a character specifying a REST API method}

\item{param}{a character specifying parameters which correspond to the selected method}

\item{user_token}{a character of a \code{oauth_token} for a user (\code{default = NULL})}

\item{user_secret}{a character of a \code{oauth_secret} for a user (\code{default = NULL})}
}
\value{
a \code{data.frame}
}
\description{
fatsecretR REST API

Main worker method for all REST API calls
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

