% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccToken.R
\name{getAccToken}
\alias{getAccToken}
\title{Get Access Token (3 Legged - Step 3)}
\usage{
getAccToken(verifierCode, userToken, tokenSecret)
}
\arguments{
\item{verifierCode}{the numeric verification code, from user authorization}

\item{userToken}{a character string for the user \code{oauth_token}}

\item{tokenSecret}{a character string for the user \code{oauth_secret}}
}
\value{
a \code{data.frame} containing the authorized \code{ouath_token} and \code{oauth_secret} for the request.
}
\description{
Retrive the \code{ouath_token} and \code{oauth_secret} from Step 3 of 3 Legged OAuth1.0 Authentication
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

