% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuth.R
\name{getAuth}
\alias{getAuth}
\title{getAuth}
\usage{
getAuth(user_id)
}
\arguments{
\item{user}{The user_id for the profile to query}
}
\value{
A \code{data.frame} containing the \code{ouath_token} and \code{oauth_secret} for the specified \code{user_id}
}
\description{
Retrive the \code{ouath_token} and \code{oauth_secret} for a profile created using \code{makeProfile}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

