% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFood.R
\name{getFood}
\alias{getFood}
\title{Query by Food item}
\usage{
getFood(CONSUMER_KEY, SHARED_SECRET, food)
}
\arguments{
\item{food}{A string of the food item to search}
}
\value{
A \code{data.frame} of food items and their corresponding ID's.
}
\description{
Query the FatSecret database for a specific food item. The main use of this function, is to to retrieve \code{food_ids} of
specific foods, which can then be passed back to the API to retrive compostional information.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

