% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoodEntry.R
\name{getFoodEntry}
\alias{getFoodEntry}
\title{get Food Entry}
\usage{
getFoodEntry(user_token, user_secret, date)
}
\arguments{
\item{user_token}{the \code{ouath_token} for the user}

\item{user_secret}{the \code{oauth_secret} for the user}

\item{date}{the date to query. The date must be in the format \emph{YYYY-MM-DD}}
}
\value{
a list of \code{data.frame's}. Each food entry is a separate element in the list; with the following \code{data.frame}
cotaining the serving size, description information and the nutritional breakdown of the food entry.
}
\description{
Retrieve all food diary entries on a given date for a specified user
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

