% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoodEntryMonth.R
\name{getFoodEntryMonth}
\alias{getFoodEntryMonth}
\title{get Food Entry Month}
\usage{
getFoodEntryMonth(user_token, user_secret, date)
}
\arguments{
\item{user_token}{the \code{ouath_token} for the user}

\item{user_secret}{the \code{oauth_secret} for the user}

\item{date}{the date to query. The date must be in the format \emph{YYYY-MM-DD}}
}
\value{
a \code{data.frame} with calories, carbohydrate, protein and fat for each day in the month that there has
been a food entry
}
\description{
Returns summary daily nutritional information for a user's food diary entries for the month specified.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

