% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoodID.R
\name{getFoodID}
\alias{getFoodID}
\title{Query database by Food ID}
\usage{
getFoodID(food_id)
}
\arguments{
\item{food_id}{a numeric value which corresponds to a valid FatSecret database entry}
}
\value{
a matrix of nutritonal information for all of the available serving sizes
}
\description{
Query the FatSecret database for a specific food item using the database ID.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

