% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoodParse.R
\name{getFoodParse}
\alias{getFoodParse}
\title{getFoodParse}
\usage{
getFoodParse(URLreq)
}
\arguments{
\item{URLreq}{the XML output from a \code{getFoodID} http GET requst}
}
\value{
a parsed XML file
}
\description{
Parse the \code{getFoodID} \code{http GET} request from \code{XML} into a readable matrix
}
\author{
Tom Wilson \code{tpw2@aber.ac.uk}
}

