% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfile.R
\name{getProfile}
\alias{getProfile}
\title{Retrieve profile information}
\usage{
getProfile(user_id)
}
\arguments{
\item{user_id}{the profile id used during \code{makeProfile}}
}
\value{
the last recorded weight measure and current height measure
}
\description{
Retrieve profile information
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

