% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReqToken.R
\name{getReqToken}
\alias{getReqToken}
\title{Get Request Token (3 Legged - Step 1)}
\usage{
getReqToken()
}
\value{
a \code{data.frame} containing the \code{ouath_token} and \code{oauth_secret} for the request
}
\description{
Retrive the \code{ouath_token} and \code{oauth_secret} from Step 1 of 3 Legged OAuth1.0 Authentication
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

