% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserProfile.R
\name{getUserProfile}
\alias{getUserProfile}
\title{User Profile}
\usage{
getUserProfile(user_token, user_secret)
}
\arguments{
\item{user_token}{the \code{ouath_token} for the user}

\item{user_secret}{the \code{oauth_secret} for the user}
}
\value{
a \code{data.frame} containing values for current weight and height measurments
}
\description{
Retrieve profile information
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

