% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeProfile.R
\name{makeProfile}
\alias{makeProfile}
\title{Create a user profile}
\usage{
makeProfile(user_id)
}
\arguments{
\item{user_id}{a unique identifier for the new profile}
}
\value{
a \code{data.frame} cotnaining the user token and access secret for the created profile
}
\description{
Create a user profile which a specific user ID and retrieve the \code{oauth_token} and \code{access_secret}
for the user. The tokens do not need to be saved, as providing a user_id string is specified,
then tokens can be retrieved at any time using \code{getAuth}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

