% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAuth.R
\name{newAuth}
\alias{newAuth}
\title{New Authorised User}
\usage{
newAuth(name = "")
}
\arguments{
\item{name}{a character of the object name}
}
\description{
Create an object for 3-Legged Authentication of an existing FatSecret user. The object is exported into
the global environment.
}
\examples{
\dontrun{
newAuth(SteveFrench)
}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
