% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/TimeStamp-method.R
\docType{methods}
\name{TimeStamp}
\alias{TimeStamp}
\alias{TimeStamp,fatsecret-method}
\alias{TimeStamp,fatsecret3L-method}
\title{oauth_timestamp}
\usage{
TimeStamp(object)

\S4method{TimeStamp}{fatsecret}(object)

\S4method{TimeStamp}{fatsecret3L}(object)
}
\arguments{
\item{object}{a \code{fatsecret} or \code{fatsecret3L} object}
}
\value{
a \code{fatsecret} or \code{fatsecret3L} object
}
\description{
oauth_timestamp

Assign a \code{oauth_timestamp} parameter to the current \code{fatsecret} or \code{fatsecret3L}
object

oauth_timestamp
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
