% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXML.R
\name{parseXML}
\alias{parseXML}
\title{Parse XML}
\usage{
parseXML(URLreq)
}
\arguments{
\item{URLreq}{the XML outout from a http GET requst}
}
\value{
parsed XML file
}
\description{
Parse the \code{http GET} request from \code{XML} into a readable matrix
}
\author{
Tom Wilson \code{tpw2@aber.ac.uk}
}

