% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXML2.R
\name{parseXML2}
\alias{parseXML2}
\title{Parse XML}
\usage{
parseXML2(URLreq)
}
\arguments{
\item{URLreq}{The XML outout from a \code{getFoodID} http GET requst}
}
\value{
Parsed XML file
}
\description{
Parse the \code{getFoodID} \code{http GET} request from \code{XML} into a readable matrix
}
\author{
Tom Wilson \code{tpw2@aber.ac.uk}
}

