% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSIXdays_to_date.R
\name{POSIXdays_to_date}
\alias{POSIXdays_to_date}
\title{Convert POSIX days to date}
\usage{
POSIXdays_to_date(x)
}
\arguments{
\item{x}{a numeric vector of \code{POSIX} days}
}
\value{
a character of date in the format \code{YYYY-MM-DD}
}
\description{
Date outputs from signed delegated requests are retured as POSIX days (number of days since \code{01/01/1970}). This function
converts the date output to a human readable format; \code{YYYY-MM-DD}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}

