% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNDSTR.R
\name{RNDSTR}
\alias{RNDSTR}
\title{Random String}
\usage{
RNDSTR(len)
}
\arguments{
\item{len}{a numeric value for the length of the string}
}
\value{
a random string of length \code{len}
}
\description{
Create a random alphanumeric string which can be used as the \code{oauth_nonce} parameter during
\code{Oauth Core 1.0} authentication
}
\examples{
\dontrun{
a_random_string <- RNDSTR(10)
print(a_random_string)
> "DJoyME7yaH"
}
}
\author{
Tom Wislon \email{tpw2@aber.ac.uk}
}

