% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/root_base_string.R
\name{root_base_string}
\alias{root_base_string}
\title{Make Root Base String}
\usage{
root_base_string(CONSUMER_KEY = CONSUMER_KEY)
}
\arguments{
\item{CONSUMER_KEY}{This is your personal alphanumeric REST API key}
}
\value{
A list of six elements
\itemize{
 \item{url} String of the HTTP method + the \code{RFC 3986} encoded API URL.
 \item{con_key} String of the user's \code{API Consumer Key}
 \item{sig_meth} Srting of the signature method
 \item{time_stamp} An interger
 \item{nonce} A string
 \item{version} A string
}
}
\description{
Make the \code{root} of the requried signature base string. In this instance,  \code{root} referes to the required parameters
which are consisent for each API query. For example, \code{oauth_timestamp},\code{oauth_nonce},\code{oauth_consumer_key}, will
always be required, thus the \code{root} base string can be created independant of query paramerts and \code{oauth_signature}.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

