% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_base_string3L.R
\name{root_base_string3L}
\alias{root_base_string3L}
\title{Make Root Base String for 3-Legged Authentication}
\usage{
root_base_string3L(CONSUMER_KEY = getOption("CONSUMER_KEY"), url, params)
}
\arguments{
\item{CONSUMER_KEY}{This is your personal alphanumeric REST API key}

\item{url}{the \code{API URL} for the specified request}

\item{params}{additional paramaters which are not included in non-signed requests}
}
\value{
a list of six elements
\itemize{
 \item{url} string of the HTTP method + the \code{RFC 3986} encoded API URL.
 \item{con_key} string of the user's \code{API Consumer Key}
 \item{sig_meth} srting of the signature method
 \item{time_stamp} an interger
 \item{nonce} a string
 \item{version} a string
 \item{callback} a string (\code{oob}) indicating \code{out-of-band configuration} callback
}
}
\description{
Make the \code{root} of the requried signature base string. In this instance,  \code{root} referes to the required parameters
which are consisent for each API query. For example, \code{oauth_timestamp},\code{oauth_nonce},\code{oauth_consumer_key}, will
always be required, thus the \code{root} base string can be created independant of query paramerts and \code{oauth_signature}.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}

