% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,fatsecret-method}
\alias{show,fatsecret-method}
\title{show-fatsecret}
\usage{
\S4method{show}{fatsecret}(object)
}
\arguments{
\item{object}{a fatsecret object}
}
\description{
show-fatsecret
}
\author{
Tom Wilsn \email{tpw2@aber.ac.uk}
}
