% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigantureValue.R
\name{signatureValue}
\alias{signatureValue}
\title{Authentication Signature Value}
\usage{
signatureValue(SIG_BASE_STR, SHARED_SECRET = getOption("SHARED_SECRET"))
}
\arguments{
\item{SIG_BASE_STR}{asignature base string which has been \code{RFC 3968} encoded.}

\item{SHARED_SECRET}{a alphanumeric string of your REST API Shared Secret.}
}
\value{
a \code{HMAC SHA1} encoded \code{oauth_signature}
}
\description{
Create the signature value which encodes the signature base string and forms the \code{oauth_signature}
parameter
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}

