% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigantureValue2.R
\name{signatureValue2}
\alias{signatureValue2}
\title{Authentication Signature Value using a User Secret}
\usage{
signatureValue2(SIG_BASE_STR, SHARED_SECRET = getOption("SHARED_SECRET"),
  ACCESS_SECRET = getOption("user_secret"))
}
\arguments{
\item{SIG_BASE_STR}{A signature base string which has been \code{RFC 3968} encoded.}

\item{SHARED_SECRET}{A alphanumeric string of your REST API Shared Secret.}

\item{ACCESS_SECRET}{A alphanumeric string of a Access Secret Token.}
}
\value{
A \code{HMAC SHA1} encoded \code{oauth_signature}
}
\description{
Create the signature value which encodes the signature base string and forms the \code{oauth_signature}
parameter. \code{signatureValue2} should be used when a \code{oauth_token} has been passed in the signature base string,
and thus the signature value must be encoded with the concatenated shared secret and user secret.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

