% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMESTAMP.R
\name{TIMESTAMP}
\alias{TIMESTAMP}
\title{Time Stamp}
\usage{
TIMESTAMP()
}
\value{
the current time and date in number of seconds since \code{January 1, 1970 00:00:00 GMT}
}
\description{
Create a POSIX timestamp which is used as the \code{oauth_timestamp} parameter during
\code{Oauth Core 1.0} authentication
}
\examples{
\dontrun{
a_time_stamp <- TIMESTAMP()
print(Sys.time())
> "2016-01-07 10:22:01 GMT"
print(a_time_stamp)
> 1452162121

}

}
\author{
Tom Wilson \email{tpw@aber.ac.uk}
}
\keyword{internal}

