% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlParser.R
\name{xmlParser}
\alias{xmlParser}
\title{Parse XML result}
\usage{
xmlParser(URLresult, method)
}
\arguments{
\item{URLresult}{the character output of \code{RCurl::getURLContent}}

\item{method}{the REST API method used}
}
\value{
a \code{data.frame} of parsed results
}
\description{
Parse the XML result of a successful REST API request
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}#'
}
