% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/APIkeys-method.R
\docType{methods}
\name{APIkeys}
\alias{APIkeys}
\alias{APIkeys,fatsecret-method}
\alias{APIkeys,fatsecret3L-method}
\title{Add API Keys}
\usage{
APIkeys(object, ConsumerKey, SharedSecret)

\S4method{APIkeys}{fatsecret}(object, ConsumerKey, SharedSecret)

\S4method{APIkeys}{fatsecret3L}(object, ConsumerKey, SharedSecret)
}
\arguments{
\item{object}{a \code{fatsecret} or \code{fatsecret3L} object}

\item{ConsumerKey}{a character of your FatSecret REST API Consumer Key}

\item{SharedSecret}{a character of your FatSecret REST API Shared Secret}
}
\value{
a \code{fatsecret} or \code{fatsecret3L} object
}
\description{
Add API Keys

Add your private API keys to the \code{fatsecret} object

Add your private API keys to the \code{fatsecret3L} object
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

