% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/fatsecretR-method.R
\docType{methods}
\name{fatsecretR}
\alias{fatsecretR}
\alias{fatsecretR,fatsecret-method}
\title{fatsecretR REST API}
\usage{
fatsecretR(object, method, params)

\S4method{fatsecretR}{fatsecret}(object, method, params)
}
\arguments{
\item{object}{a \code{fatsecret} object}

\item{method}{a character specifying a REST API method}

\item{params}{a character specifying parameters which correspond to the selected method}
}
\value{
a \code{data.frame}
}
\description{
fatsecretR REST API

Main worker method for all REST API calls
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

