% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoodEntryMonth.R
\name{getFoodEntryMonth}
\alias{getFoodEntryMonth}
\title{get Food Entry Month}
\usage{
getFoodEntryMonth(user_token, user_secret, date)
}
\arguments{
\item{user_secret}{the \code{oauth_secret} for the user}

\item{date}{the date to query. The date must be in the format \emph{YYYY-MM-DD}}

\item{user_toker}{the \code{ouath_token} for the user}
}
\value{
a list of two elements
 \itemize{
   \item {dates} the start and end date of the search
   \item {summary} a \code{data.frame} summary of nutritional intake for the month
 }
}
\description{
Returns summary daily nutritional information for a user's food diary entries for the month specified.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

