% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoodID.R
\name{getFoodID}
\alias{getFoodID}
\title{Query by Food ID}
\usage{
getFoodID(food_id)
}
\arguments{
\item{food_id}{A numeric value which corresponds to a valid FatSecret database entry}
}
\value{
A list of two elements
 \itemize{
   \item{id} The food descriptor information
   \item{servings} The nutritional breakdown of the food by servings
 }
}
\description{
Query the FatSecret database for a specific food item using the database ID.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

