% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRequest.R
\name{makeRequest}
\alias{makeRequest}
\title{Get Request Token (3 Legged - Step 1)}
\usage{
makeRequest(callback)
}
\arguments{
\item{callback}{a charatcer string for the `oauth_callback`. If unable to recieve a callback, then `callback` must be
set to `oob`, otherwise `callback` should equal the absolute url of your callback}
}
\value{
a \code{data.frame} containing the following four values;
 \itemize{
   \item callback
   \item request_token
   \item token_secret
   \item request_url
 }
}
\description{
Retrive the \code{ouath_token} and \code{oauth_secret} from Step 1 of 3 Legged OAuth1.0 Authentication
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

