% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/OauthNonce-method.R
\docType{methods}
\name{OauthNonce}
\alias{OauthNonce}
\alias{OauthNonce,fatsecret-method}
\alias{OauthNonce,fatsecret3L-method}
\title{oauth_nonce}
\usage{
OauthNonce(object)

\S4method{OauthNonce}{fatsecret}(object)

\S4method{OauthNonce}{fatsecret3L}(object)
}
\arguments{
\item{object}{a \code{fatsecret} or \code{fatsecret3L} object}
}
\value{
a \code{fatsecret} or \code{fatsecret3L} object
}
\description{
oauth_nonce

Assign a \code{oauth_nonce} parameter to the current \code{fatsecret} or \code{fatsecret3L}
object

oauth_nonce
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
