% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigantureValue.R
\name{signatureValue}
\alias{signatureValue}
\title{Authentication Signature Value}
\usage{
signatureValue(SIG_BASE_STR, SHARED_SECRET, ACCESS_SECRET = NULL)
}
\arguments{
\item{SIG_BASE_STR}{A signature base string which has been \code{RFC 3968} encoded.}

\item{SHARED_SECRET}{A alphanumeric string of your REST API Shared Secret.}

\item{ACCESS_SECRET}{A alphanumeric string of a Access Secret Token. Default is \code{NULL}.}
}
\value{
A \code{HMAC SHA1} encoded \code{oauth_signature}
}
\description{
Create the signature value which encodes the signature base string and forms the \code{oauth_signature}
parameter
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

